+++ e2e/__tests__/jasmineAsync.test.ts
@@ -5,6 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import {isJestJasmineRun} from '@jest/test-utils';
 import runJest, {json as runWithJson} from '../runJest';
 
 describe('async jasmine', () => {
@@ -107,16 +108,25 @@ describe('async jasmine', () => {
   });
 
   it('works with concurrent', () => {
-    const {json} = runWithJson('jasmine-async', ['concurrent.test.js']);
+    const {json, stderr} = runWithJson('jasmine-async', ['concurrent.test.js']);
     expect(json.numTotalTests).toBe(4);
     expect(json.numPassedTests).toBe(2);
     expect(json.numFailedTests).toBe(1);
     expect(json.numPendingTests).toBe(1);
     expect(json.testResults[0].message).toMatch(/concurrent test fails/);
+    if (!isJestJasmineRun()) {
+      expect(stderr.match(/\[\[\w+\]\]/g)).toEqual([
+        '[[beforeAll]]',
+        '[[test]]',
+        '[[test]]',
+        '[[test]]',
+        '[[afterAll]]',
+      ]);
+    }
   });
 
   it('works with concurrent within a describe block when invoked with testNamePattern', () => {
-    const {json} = runWithJson('jasmine-async', [
+    const {json, stderr} = runWithJson('jasmine-async', [
       '--testNamePattern',
       'one concurrent test fails',
       'concurrentWithinDescribe.test.js',
@@ -126,6 +136,8 @@ describe('async jasmine', () => {
     expect(json.numFailedTests).toBe(1);
     expect(json.numPendingTests).toBe(1);
     expect(json.testResults[0].message).toMatch(/concurrent test fails/);
+    expect(stderr).toMatch(/this is logged \d/);
+    expect(stderr).not.toMatch(/this is not logged \d/);
   });
 
   it('works with concurrent.each', () => {